Variables are initially being set so you can refer to them later. Then, a function is created to generate a list item and change the text inside it.
Lastly, another function is established to grab input from the user and insert it into a variable. After that, the function that creates the list item is called, inserting the user input text into the previously defined variable. This process adds the newly created list item into the unordered list element in the HTML.

Finally, the input field is cleared, resetting its value to zero for fresh new input. An event listener for clicks is added to the Add button, allowing these functions to begin.